+++ src/features/auth/components/LoginForm.tsx
@@ -12,6 +12,7 @@ import {
 } from "@/shared/utils/validation";
 import Skeleton from "@/components/Skeleton";
 import { toast } from "react-toastify";
+import { toastApiError } from "@/shared/utils/toastApiError";
 import { ROUTES } from "@/shared/constants/routes";
 
 export default function LoginForm() {
@@ -58,7 +59,7 @@ export default function LoginForm() {
       const message =
         (error as Error)?.message ?? "잠시 후 다시 시도해 주세요.";
       setSubmitError(message);
-      toast.error("잠시 후 다시 시도해 주세요.");
+      toastApiError(error);
     } finally {
       setIsSubmitting(false);
     }
