+++ src/shared/utils/toastApiError.ts
@@ -1,0 +1,6 @@
+import { toast } from "react-toastify";
+import { normalizeError } from "@/shared/lib/http";
+
+export const toastApiError = (error: unknown) => {
+  toast.error(normalizeError(error).message);
+};
